Support transport mode
If a policy (configuration) does not specify any peers,
then flag the policy as being able to use transport mode
IPSec. Send the USE_TRANSPORT_MODE notification with the
request to signal the peer that we want transport mode.
Enable transport mode if we receive the USE_TRANSPORT_MODE
notification.